api-gateway/requirements.txt
blinker
certifi
cffi
charset-normalizer
click
cryptography
Flask
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
requests
urllib3
Werkzeug
PyJWT